+++ src/lib/sparql_shacl.py
@@ -42,28 +42,30 @@ def get_shacl_properties() -> OntologyProperty:
 
     text = """
         SELECT
-            ?uri
+            (COALESCE(?target_class_, '') as ?card_of_class_uri)
             (COALESCE(?label_, ?uri) as ?label)
             (COALESCE(?order_, '') as ?order)
             (COALESCE(?min_count_, '') as ?min_count)
             (COALESCE(?max_count_, '') as ?max_count)
-            ?domain_class_uri
-            (COALESCE(?range_class_uri_1, ?range_class_uri_2, '') as ?range_class_uri)
-            (isBlank(?uri) as ?is_blank)
+            (COALESCE(?domain_class_uri_, '') as ?domain_class_uri)
+            ?uri
+            (COALESCE(?range_class_uri_, ?datatype_, '') as ?range_class_uri)
         WHERE {
-            """ + ("GRAPH " + ensure_uri(endpoint.ontology_uri) + " {" if endpoint.ontology_uri else "") + """
+            """ + ("GRAPH " + ensure_uri(endpoint.ontology_uri) + " {" if endpoint.ontology_uri else "") + """               
                 ?shape sh:property ?node .
-                ?node sh:path ?uri .  
-                ?shape sh:targetClass ?domain_class_uri .
+                ?node sh:path ?supposed_uri .  
+                OPTIONAL { ?shape sh:targetClass ?target_class_ . }
+                OPTIONAL { ?supposed_uri sh:inversePath ?inverse_property_uri . }
                 OPTIONAL { ?node sh:name ?label_ . }
                 OPTIONAL { ?node sh:order ?order_ . }
                 OPTIONAL { ?node sh:minCount ?min_count_ . }
                 OPTIONAL { ?node sh:maxCount ?max_count_ . }
-                OPTIONAL {
-                    ?node sh:class ?range_class_uri_1 .
-                    ?range_shape sh:targetClass ?range_class_uri_1 .
-                }
-                OPTIONAL { ?node sh:datatype ?range_class_uri_2 . }
+                OPTIONAL { ?node sh:datatype ?datatype_ . }
+                OPTIONAL { ?node sh:class ?class . }
+
+                BIND(IF(isBlank(?supposed_uri), '', ?target_class_) as ?domain_class_uri_)
+                BIND(IF(isBlank(?supposed_uri), ?target_class_, ?class) as ?range_class_uri_)
+                BIND(IF(isBlank(?supposed_uri), ?inverse_property_uri, ?supposed_uri) as ?uri)
             """ + ("}" if endpoint.ontology_uri else "") + """
         }
     """