+++ src/pages/entity.py
@@ -69,8 +69,8 @@ elif not entity:
 else:
 
     # Header: entity name, additional info and description
-    col1, col2, col_delete, col_edit = st.columns([5, 1, 1, 1], vertical_alignment='bottom')
-    col1.title(entity.display_label)
+    col1, col2, col_delete, col_edit = st.columns([20, 1, 1, 4], vertical_alignment='bottom')
+    col1.markdown(f'# {entity.display_label} <small style="font-size: 16px; color: gray; text-decoration: none;">{entity.uri}</small>', unsafe_allow_html=True)
     col2.button('', icon=':material/info:', type='tertiary', on_click=dialog_entity_info, kwargs={'entity': entity})
     if col_delete.button('', icon=':material/delete:', type='primary'):
         dialog_confirmation("You are about to delete this entity (including all triples) in this graph.", __delete_entity, entity=entity, graph=graph)